Moves table commands to tables.scala
